DT-337 Use runtime exception where single active custody event expected
By moving the exception handling into SecureControllerAdvice we lose the
ability to have bespoke error messages returned to the user.  However
we remove lots of exception handling from the main code and this was
considered to be a reasonable trade off.

Also the error messages were previously attached to
ErrorResponse#userMessage but the new pattern in SecureControllerAdvice
attaches them to ErrorResponse#developerMessage so I've gone with that.
As these error responses are not being consumed by another system yet
this should not be a problem.