DCS-309 Adding endpoint to allow setting roles
Adding new endpoint to allow idempotent assigning of roles.

Hit a limitation of the embedded ldap server which doesn't support aliases which is how delius model roles.
So locally we cannot retrieve descriptions of roles - we don't think descriptions are used by any clients so removing.

Also moving from using uid as part of a user's dn to use cn instead. This is consistent with the delius codebase and we
think it should lead to equivalent behaviour.

Also moving OffenderResource to have a consistent swagger description as the other secure endpoints - as it's currently hard to see in the docs.